docker/athanor-ingest/requirements.txt
chromadb==1.5.7
google-cloud-storage==3.10.1
google-cloud-secret-manager==2.22.0
google-auth==2.49.1
httpx==0.28.1
openai==2.31.0
PyMuPDF==1.27.2.2
python-docx==1.2.0
python-pptx==1.0.2
networkx==3.6.1
pyyaml==6.0.2